feat: Added Virtual Machines
This commit introduces a new directory structure for managing virtual
machines, specifically tailored for setting up Kubernetes manually on
Ubuntu. Files have been added to the `virtual-machines` directory,
which includes scripts and configurations necessary for the setup.
The following files have been added:
- `README.md`: Documentation for the virtual machines setup.
- `Vagrantfile`: Configuration for Vagrant to manage the virtual machines.
- `ssh.sh`: Script to set up SSH access.
- `tmux.conf`: Configuration file for tmux.
- `update-dns.sh`: Script to update DNS settings.
- `controlplane.sh`: Script to set up the control plane.
- `setup-hosts.sh`: Script to configure the hosts file.
- `vimrc`: Configuration file for Vim.

This commit is part of the ongoing effort to enhance the documentation
and setup instructions for manual installation of Kubernetes. Making
it easier for users to get started with manual installations and
configurations. With the intended purpose of making them more capabale
administrators and users of Kubernetes.